requirements.txt
torch
fastapi
pillow
numpy
